fix all MVC names and separate the post booking method
