+++ static/booking.html
@@ -273,5 +273,8 @@
   </body>
   <!-- <script type="module" src="/static/js/auth.js"></script> -->
   <script type="module" src="/static/js/booking.js"></script>
-  <script type="module" src="/static/js/controller/booking.js"></script>
+  <script
+    type="module"
+    src="/static/js/controller/controller_booking.js"
+  ></script>
 </html>
