+++ static/js/auth.js
@@ -1,4 +1,4 @@
-import { initialize, checkUserState } from "./controller/auth.js";
+import { initialize, checkUserState } from "./controller/controller_auth.js";
 
 const signinMask = document.querySelector(".signin-mask");
 const signupMask = document.querySelector(".signup-mask");
