+++ static/js/controller/controller_thankyou.js
