+++ static/js/view/view_attraction.js
