+++ static/js/view/view_booking.js
