+++ static/thankyou.html
@@ -124,5 +124,8 @@
   </body>
   <script type="module" src="/static/js/auth.js"></script>
   <script type="module" src="/static/js/booking.js"></script>
-  <script type="module" src="/static/js//controller/thankyou.js"></script>
+  <script
+    type="module"
+    src="/static/js//controller/controller_thankyou.js"
+  ></script>
 </html>
